LANEWEB-10545: upgrade embedded tomcat
tomcat-9.0.31 changed default AJP connector values, including:

secretRequired

If this attribute is true, the AJP Connector will only start if the
secret attribute is configured with a non-null, non-zero length value.
The default value is true. This attributue should only be set to false
when the Connector is used on a trusted network.