LANEWEB-10906: staff slider size/font adjustments
"reduce the library staff from 6 sections to 5 so we have more space for
the name and job desc. Please view my screenshot below. Also, one thing
I noticed is that the job desc font is thin so I suggest: font-weight:
300;

Also, we have some white space on both sides of the "Meet Your Lane
Library Staff" that we can use to make the image a little bit wider so
there is enough space for the job desc text so in this case, we might
not even need to remove/reduce the sections from 6 to 5..."